requirements.txt
# Google GenAI SDK (unified SDK replacing deprecated google-generativeai)
# Migration: https://ai.google.dev/gemini-api/docs/migrate
google-genai>=1.0.0

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Configuration
python-dotenv==1.0.1
pydantic>=2.11.0

# Async Support
aiohttp==3.9.3

# Rate Limiting
slowapi==0.1.9

# Optional: Google Cloud Observability (uncomment for Cloud Run)
# opentelemetry-api>=1.20.0
# opentelemetry-sdk>=1.20.0
# opentelemetry-exporter-gcp-trace>=1.5.0

# Braintrust Evals (pip install braintrust[performance] for faster JSON)
braintrust>=0.4.0
autoevals>=0.0.52

# Evals Dependencies
pyyaml>=6.0
httpx>=0.24.0
